+++ src/workspace-tree/bazel_workspace_folder_tree_item.ts
@@ -160,7 +160,11 @@ export class BazelWorkspaceFolderTreeItem implements IBazelTreeItem {
     const packagePaths = await new BazelQuery(
       getDefaultBazelExecutablePath(),
       workspacePath,
-    ).queryPackages("...:*");
+    ).queryPackages(
+      vscode.workspace
+        .getConfiguration("bazel.commandLine")
+        .get("queryExpression"),
+    );
     const topLevelItems: BazelPackageTreeItem[] = [];
     this.buildPackageTree(
       packagePaths,